Added support for REST API with JWT.

+++ src/main/java/com/adm/dao/IDao.java
@@ -6,7 +6,7 @@ public interface IDao<T> {
 
     public List<T> getList();
 
-    public T findById(Long id);
+    public T findById(Integer id);
 
     public void update(T entity);
